AD-126 Split DataAccess into GraphQLDataClient and RESTDataClient
